- version 2.0.0

+++ RemoveBOM-VS/MainForm.cs
@@ -11,6 +11,8 @@ namespace RemoveBOM
 {
     public partial class MainForm : Form
     {
+        private int countBOM = 0;
+
         public MainForm()
         {
             InitializeComponent();
@@ -26,17 +28,25 @@ namespace RemoveBOM
 
         private void listFiles_DragDrop(object sender, DragEventArgs e)
         {
+            this.countBOM = 0;
+
             string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
 
             foreach (string file in files)
             {
                 applyTo(file);
             }
+
+            MessageBox.Show("Files with BOM: " + this.countBOM.ToString(), "BOM", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            this.lblBOMFiles.Text = this.countBOM.ToString();
         }
 
         private void readDirectory(string dirPath)
         {
-            this.listFiles.Items.Add(dirPath + " [DIR]");
+            if (this.rbAll.Checked)
+            {
+                this.listFiles.Items.Add(dirPath + " [DIR]");
+            }
 
             string[] files = Directory.GetFiles(dirPath);
 
@@ -61,21 +71,31 @@ namespace RemoveBOM
             }
             else if (File.Exists(file))
             {
-                this.listFiles.Items.Add(file + " [" + removeBOMFromFile(file) + "]");
+                if (processBOMFromFile(file))
+                {
+                    this.listFiles.Items.Add(file + " [" + (this.rbRemove.Checked ? "REMOVE" : "INCLUDES") + " BOM]");
+                }
+                else if(this.rbAll.Checked)
+                {
+                    this.listFiles.Items.Add(file + " [NO BOM]");
+                }
             }
             else
             {
-                this.listFiles.Items.Add(file + " [ERROR]");
+                if(this.rbAll.Checked)
+                {
+                    this.listFiles.Items.Add(file + " [ERROR]");
+                }
             }
         }
 
-        private string removeBOMFromFile(string file)
+        private bool processBOMFromFile(string file)
         {
             FileStream read = File.OpenRead(file);
 
             byte[] bufferBOM = new byte[3];
 
-            string info = "";
+            bool bom = false;
 
             int bomCount = read.Read(bufferBOM, 0, bufferBOM.Length);
 
@@ -83,50 +103,51 @@ namespace RemoveBOM
             {
                 if ((bufferBOM[0] == 239) && (bufferBOM[1] == 187) && (bufferBOM[2] == 191))
                 {
-                    byte[] buffer = new byte[1024];
+                    this.countBOM++;
 
-                    FileStream write = File.OpenWrite(file + ".utf8");
-
-                    int readCount = read.Read(buffer, 0, buffer.Length); ;
-
-                    while (readCount > 0)
+                    if (this.rbRemove.Checked)
                     {
-                        write.Write(buffer, 0, readCount);
-                        readCount = read.Read(buffer, 0, buffer.Length);
+                        byte[] buffer = new byte[1024];
+
+                        FileStream write = File.OpenWrite(file + ".utf8");
+
+                        int readCount = read.Read(buffer, 0, buffer.Length); ;
+
+                        while (readCount > 0)
+                        {
+                            write.Write(buffer, 0, readCount);
+                            readCount = read.Read(buffer, 0, buffer.Length);
+                        }
+
+                        write.Close();
+
+                        read.Close();
+
+                        if (this.chkBackup.Checked)
+                        {
+                            System.IO.File.Move(file, file + ".bak");
+                        }
+                        else
+                        {
+                            System.IO.File.Delete(file);
+                        }
+
+                        System.IO.File.Move(file + ".utf8", file);
                     }
 
-                    write.Close();
-
-                    read.Close();
-                    
-                    if (this.chkBackup.Checked)
-                    {
-                        System.IO.File.Move(file, file + ".bak");
-                    }
-                    else
-                    {
-                        System.IO.File.Delete(file);
-                    }
-
-                    System.IO.File.Move(file + ".utf8", file);
-
-                    info = "BOM";
+                    bom = true;
                 }
                 else
                 {
                     read.Close();
-
-                    info = "NO BOM";
                 }
             }
             else
             {
                 read.Close();
-
-                info = "NO BOM";
             }
 
-            return info;
+            return bom;
         }
 
         private void lnkForrest79_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
@@ -134,9 +155,10 @@ namespace RemoveBOM
             System.Diagnostics.Process.Start("http://forrest79.net");
         }
 
-        private void btnClear_Click(object sender, EventArgs e)
+        private void btnClearList_Click(object sender, EventArgs e)
         {
             this.listFiles.Items.Clear();
+            this.lblBOMFiles.Text = "0";
         }
     }
 }
